+++ src/sll/api/_generated_raw.c
@@ -181,23 +181,21 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_audit_audit_raw(sll_obje
 
 
 
-__SLL_API_TYPE_sll_api_audit_register_callback sll_api_audit_register_callback(__SLL_API_ARGS_sll_api_audit_register_callback);
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_audit_register_callback_raw(sll_object_t*const* al,sll_arg_count_t all){
-	sll_integer_t* a;
-	sll_arg_count_t ac;
-	sll_parse_args(SLL_CHAR("i+"),al,all,&a,&ac);
-	sll_api_audit_register_callback(a,ac);
+__SLL_API_TYPE_sll_api_audit_register sll_api_audit_register(__SLL_API_ARGS_sll_api_audit_register);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_audit_register_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_integer_t a;
+	sll_parse_args(SLL_CHAR("i"),al,all,&a);
+	sll_api_audit_register(a);
 	return SLL_ACQUIRE_STATIC_INT(0);
 }
 
 
 
-__SLL_API_TYPE_sll_api_audit_unregister_callback sll_api_audit_unregister_callback(__SLL_API_ARGS_sll_api_audit_unregister_callback);
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_audit_unregister_callback_raw(sll_object_t*const* al,sll_arg_count_t all){
-	sll_integer_t* a;
-	sll_arg_count_t ac;
-	sll_parse_args(SLL_CHAR("i+"),al,all,&a,&ac);
-	sll_bool_t out=sll_api_audit_unregister_callback(a,ac);
+__SLL_API_TYPE_sll_api_audit_unregister sll_api_audit_unregister(__SLL_API_ARGS_sll_api_audit_unregister);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_audit_unregister_raw(sll_object_t*const* al,sll_arg_count_t all){
+	sll_integer_t a;
+	sll_parse_args(SLL_CHAR("i"),al,all,&a);
+	sll_bool_t out=sll_api_audit_unregister(a);
 	SLL_ACQUIRE(sll_static_int[out]);
 	return sll_static_int[out];
 }
@@ -2190,12 +2188,12 @@ static const internal_function_t _ifunc_data_ptr[]={
 		sll_api_audit_audit_raw
 	},
 	{
-		SLL_CHAR("sll:audit_register_callback"),
-		sll_api_audit_register_callback_raw
+		SLL_CHAR("sll:audit_register"),
+		sll_api_audit_register_raw
 	},
 	{
-		SLL_CHAR("sll:audit_unregister_callback"),
-		sll_api_audit_unregister_callback_raw
+		SLL_CHAR("sll:audit_unregister"),
+		sll_api_audit_unregister_raw
 	},
 	{
 		SLL_CHAR("sll:base64_decode"),