+++ src/sll/audit.c
@@ -1,8 +1,11 @@
 #include <sll/_internal/common.h>
+#include <sll/array.h>
 #include <sll/audit.h>
 #include <sll/common.h>
 #include <sll/gc.h>
+#include <sll/memory.h>
 #include <sll/new_object.h>
+#include <sll/object.h>
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/var_arg.h>
@@ -10,7 +13,16 @@
 
 
 
+static sll_audit_callback_t* _audit_cb=NULL;
 static sll_array_length_t _audit_cb_len=0;
+
+
+
+void _audit_cleanup(void){
+	sll_deallocate(_audit_cb);
+	_audit_cb=NULL;
+	_audit_cb_len=0;
+}
 
 
 
@@ -36,10 +48,40 @@ __SLL_EXTERNAL void sll_audit_list(const sll_char_t* nm,const sll_char_t* t,sll_
 	if (!_audit_cb_len){
 		return;
 	}
-	sll_object_t* o=sll_new_object_list(t,sll_string_length_unaligned(t),va);
+	sll_string_t nm_s;
+	sll_string_from_pointer(nm,&nm_s);
+	sll_array_t arr;
+	sll_new_object_array_list(t,sll_string_length_unaligned(t),va,&arr);
 	for (sll_array_length_t i=0;i<_audit_cb_len;i++){
-		SLL_UNIMPLEMENTED();
+		(*(_audit_cb+i))(&nm_s,&arr);
 	}
-	sll_release_object(o);
+	sll_free_array(&arr);
+	sll_free_string(&nm_s);
 }
 
+
+
+__SLL_EXTERNAL void sll_audit_register_callback(sll_audit_callback_t fn){
+	_audit_cb_len++;
+	_audit_cb=sll_reallocate(_audit_cb,_audit_cb_len*sizeof(sll_audit_callback_t));
+	*(_audit_cb+_audit_cb_len-1)=fn;
+}
+
+
+
+__SLL_EXTERNAL sll_bool_t sll_audit_unregister_callback(sll_audit_callback_t fn){
+	sll_array_length_t i=0;
+	sll_array_length_t j=0;
+	for (;j<_audit_cb_len;j++){
+		if (*(_audit_cb+j)!=fn){
+			*(_audit_cb+i)=*(_audit_cb+j);
+			i++;
+		}
+	}
+	if (i==j){
+		return 0;
+	}
+	_audit_cb_len=i;
+	_audit_cb=sll_reallocate(_audit_cb,_audit_cb_len*sizeof(sll_audit_callback_t));
+	return 1;
+}