+++ src/sll/include/sll/_internal/audit.h
@@ -1,0 +1,14 @@
+#ifndef __SLL__INTERNAL_AUDIT_H__
+#define __SLL__INTERNAL_AUDIT_H__ 1
+
+
+
+void _audit_cleanup(void);
+
+
+
+void _audit_cleanup_api(void);
+
+
+
+#endif