+++ src/sll/include/sll/new_object.h
@@ -36,6 +36,20 @@ __SLL_EXTERNAL void sll_new_object_array(const sll_char_t* t,sll_array_t* o,...)
 
 
 /**
+ * \flags func
+ * \name sll_new_object_list
+ * \group new-object
+ * \desc Docs!
+ * \arg const sll_char_t* t
+ * \arg sll_string_length_t tl
+ * \arg sll_var_arg_list_t* va
+ * \arg sll_array_t* o
+ */
+__SLL_EXTERNAL void sll_new_object_array_list(const sll_char_t* t,sll_string_length_t tl,sll_var_arg_list_t* va,sll_array_t* o);
+
+
+
+/**
  * \flags check_output func
  * \name sll_new_object_list
  * \group new-object