+++ src/sll/scheduler.c
@@ -1,3 +1,4 @@
+#include <sll/_internal/audit.h>
 #include <sll/_internal/barrier.h>
 #include <sll/_internal/common.h>
 #include <sll/_internal/dispatcher.h>
@@ -175,6 +176,7 @@ sll_return_code_t _scheduler_run(void){
 		ptr=PTR(ADDR(ptr)+sz);
 	}
 	_cpu_core_worker(b_ptr);
+	_audit_cleanup_api();
 	SLL_CRITICAL(sll_platform_set_cpu(SLL_CPU_ANY));
 	ptr=b_ptr;
 	for (sll_cpu_t i=0;i<_scheduler_load_balancer.len;i++){